Fix Y2038 timestamp overflow in FDB and Recorder components
- Change FdbInfo timestamp from uint32_t to uint64_t to prevent Y2038 overflow
- Update SwitchStateBase::processFdbEntriesForAging() to use uint64_t for current time
- Update SwitchStateBaseFdb::process_packet_for_fdb_event() to use uint64_t for frametime
- Replace gettimeofday() with clock_gettime() in Recorder::getTimestamp() for Y2038 safety
- Add <ctime> header for clock_gettime support

+++ vslib/SwitchStateBaseFdb.cpp
@@ -395,7 +395,8 @@ void SwitchStateBase::process_packet_for_fdb_event(
     // we would need hostif info here and maybe interface index, then we can
     // find host info from index
 
-    uint32_t frametime = (uint32_t)time(NULL);
+    // Use 64-bit timestamp to avoid Y2038 overflow
+    uint64_t frametime = static_cast<uint64_t>(time(NULL));
 
     /*
      * We add +2 in case if frame contains 1Q VLAN tag.
